requirements.txt
pytorch-lightning==1.2
transformers @ git+https://github.com/ZurichNLP/transformers.git@longmbart#egg=transformers
tensorboardX
test-tube==0.7.5
nltk==3.5
numpy
#nlp
rouge_score==0.0.4
requests
sacrebleu
datasets==1.2.1
sentencepiece
# required for datasets
dill==0.3.4
spacy